patch zone serial to use int64's instead of int32
According to RFC1035 section 3.3.13 the serial number of a SOA resource
record is a 32 bit unsigned integer. There is no current way of
specifying unsigned integers in the swagger spec, therefore let's patch
it with int64's for now while we figure out what we want to do within
the swagger spec itself.

+++ zone.go
@@ -31,10 +31,10 @@ type Zone struct {
 	Rrsets []RrSet `json:"rrsets,omitempty"`
 
 	// The SOA serial number
-	Serial int32 `json:"serial,omitempty"`
+	Serial int64 `json:"serial,omitempty"`
 
 	// The SOA serial notifications have been sent out for
-	NotifiedSerial int32 `json:"notified_serial,omitempty"`
+	NotifiedSerial int64 `json:"notified_serial,omitempty"`
 
 	//  List of IP addresses configured as a master for this zone (“Slave” type zones only)
 	Masters []string `json:"masters,omitempty"`
